Change address base type to Long
* addresses can be larger than 32bit, thus need to use Long
* in fact, having the upper-most bit 1 already gave problems
* fixed and updated all dependents

+++ src/main/scala/axi4lite/RegisterFile.scala
@@ -12,12 +12,12 @@ object RegisterFile {
    *  @param width Register data width (in bits).
    *  @param regs Map from offsets in addrGranularity to register implementations.
    **/
-  case class Configuration(addressWordBits: Int = 8, regs: Map[Int, ControlRegister], fifoDepth: PosInt = 2)
+  case class Configuration(addressWordBits: Int = 8, regs: Map[Long, ControlRegister], fifoDepth: PosInt = 2)
                           (implicit axi: Axi4Lite.Configuration) {
     /* internal helpers: */
     private def overlap(p: (BitRange, BitRange)) = p._1.overlapsWith(p._2)
-    private def makeRange(a: Int): BitRange =
-      BitRange(a * addressWordBits + axi.dataWidth.toInt - 1, a * addressWordBits)
+    private def makeRange(a: Long): BitRange =
+      BitRange(a * addressWordBits + axi.dataWidth.toLong - 1, a * addressWordBits)
     private lazy val m = regs.keys.toList.sorted map makeRange
     private lazy val o = (m.take(m.length - 1) zip m.tail) map { case (r1, r2) => ((r1, r2), r1.overlapsWith(r2)) }
     o filter (_._2) foreach { case ((r1, r2), _) => require(!r1.overlapsWith(r2), s"$r1 and $r2 must not overlap") }
